// MovieStore/movie-store-app/src/app/api/api-reference.ts
export class Movie implements IMovie {
    id?: string;
    name?: string;
    dateOfRelease?: Date | undefined;
    licensingType?: LicensingTypes;
    price?: Money | undefined;

    constructor(data?: IMovie) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.dateOfRelease = _data["dateOfRelease"] ? new Date(_data["dateOfRelease"].toString()) : <any>undefined;
            this.licensingType = _data["licensingType"];
            this.price = _data["price"] ? Money.fromJS(_data["price"]) : <any>undefined;
        }
    }

    static fromJS(data: any): Movie {
        data = typeof data === 'object' ? data : {};
        let result = new Movie();
        result.init(data)
        return result
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["dateOfRelease"] = this.dateOfRelease ? this.dateOfRelease.toISOString() : <any>undefined;
        data["licensingType"] = this.licensingType;
        data["price"] = this.price ? this.price.toJSON() : <any>undefined;
        return data;
    }
}

export interface IMovie {
    id?: string;
    name?: string;
    dateOfRelease?: Date | undefined;
    licensingType?: LicensingTypes;
    price?: Money | undefined;
}

export enum LicensingTypes {
    TwoDay = 0,
    LifeLong = 1,
}

export class Money implements IMoney {
    value?: number;

    constructor(data?: IMoney) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): Money {
        data = typeof data === 'object' ? data : {};
        let result = new Money();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["value"] = this.value;
        return data;
    }
}

export interface IMoney {
    value?: number;
}
// ...
export class CreateMovieCommand implements ICreateMovieCommand {
    name?: string;
    dateOfRelease?: Date | undefined;
    licensingType?: LicensingTypes;

    constructor(data?: ICreateMovieCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.dateOfRelease = _data["dateOfRelease"] ? new Date(_data["dateOfRelease"].toString()) : <any>undefined;
            this.licensingType = _data["licensingType"];
        }
    }

    static fromJS(data: any): CreateMovieCommand {
        data = typeof data === 'object' ? data : {};
        let result = new CreateMovieCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["dateOfRelease"] = this.dateOfRelease ? this.dateOfRelease.toISOString() : <any>undefined;
        data["licensingType"] = this.licensingType;
        return data;
    }
}

export interface ICreateMovieCommand {
    name?: string;
    dateOfRelease?: Date | undefined;
    licensingType?: LicensingTypes;
}

export class UpdateMovieCommand implements IUpdateMovieCommand {
    id?: string;
    name?: string;
    dateOfRelease?: Date | undefined;

    constructor(data?: IUpdateMovieCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.dateOfRelease = _data["dateOfRelease"] ? new Date(_data["dateOfRelease"].toString()) : <any>undefined;
        }
    }

    static fromJS(data: any): UpdateMovieCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateMovieCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["dateOfRelease"] = this.dateOfRelease ? this.dateOfRelease.toISOString() : <any>undefined;
        return data;
    }
}

export interface IUpdateMovieCommand {
    id?: string;
    name?: string;
    dateOfRelease?: Date | undefined;
}